app/utils/vector_store.py
import os

import google.auth
from langchain_community.vectorstores import FAISS
from langchain_text_splitters import CharacterTextSplitter
from langchain_google_vertexai import VertexAIEmbeddings

from app.utils.document_loader import load_documents

_, project_id = google.auth.default()
LOCATION = os.getenv("LOCATION", "europe-southwest1")

_vector_store = None


def get_embeddings() -> VertexAIEmbeddings:
    return VertexAIEmbeddings(
        model_name="textembedding-gecko@003",
        project=project_id,
        location=LOCATION,
    )


def get_vector_store():
    """Get or create vector store with documents."""
    global _vector_store
    if _vector_store is None:
        docs = load_documents()
        splitter = CharacterTextSplitter(chunk_size=1000, chunk_overlap=200)
        chunks = splitter.split_documents(docs)
        embeddings = get_embeddings()
        _vector_store = FAISS.from_documents(chunks, embeddings)
    return _vector_store

